requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flasgger==0.9.5
Flask==2.0.3
grpcio==1.44.0
grpcio-tools==1.44.0
idna==3.3
imutils==0.5.4
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
mistune==2.0.2
numpy==1.22.2
opencv-python==4.5.5.62
pconst==1.0.0
protobuf==3.19.4
pyrsistent==0.18.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
